Serial output in volts/amps, treat VLOAD as signed
- current printed in A (4 decimals), voltages in V (3 decimals)
- tsc_vload_mv now signed: a floating VLOAD input reads slightly
  below 0 V and previously showed up as ~130 V unsigned garbage

+++ Core/Src/tsc1641.c
@@ -6,7 +6,7 @@
 #include "xnx_i2c.h"
 
 volatile int32_t tsc_current_ua = 0;
-volatile uint32_t tsc_vload_mv = 0;
+volatile int32_t tsc_vload_mv = 0;
 volatile uint32_t tsc_power_mw = 0;
 volatile int16_t tsc_vshunt_raw = 0;
 volatile uint8_t tsc_online = 0;
@@ -55,7 +55,7 @@ void TSC1641_Update(void)
   if (xnx_i2c_read_reg16_ex(tsc_bus, TSC1641_ADDR, TSC1641_REG_VLOAD,
                             &raw) == HAL_OK)
   {
-    tsc_vload_mv = (uint32_t)raw * 2U;
+    tsc_vload_mv = (int32_t)(int16_t)raw * 2;
   }
   else
   {
